Fixed some code.

+++ src/game/Zone.c
@@ -77,17 +77,17 @@ void Zone(const struct AutoSpaceZone *const sz) {
 
 	/* some asteroids */
 	for(i = 0; i < /*7000*/1000; i++) {
-		s = Sprite(SP_DEBRIS, AutoImageSearch("Asteroid.png"), (int)rnd(de_sitter), (int)rnd(de_sitter), rnd((float)M_PI), big_asteroid_mass);
+		s = Sprite(SP_DEBRIS, AutoImageSearch("Asteroid.png"), rnd(de_sitter), rnd(de_sitter), rnd((float)M_PI), big_asteroid_mass);
 		SpriteSetVelocity(s, rnd(0.02f), rnd(0.02f));
 		SpriteSetOmega(s, rnd(10.0f));
 	}
 
 	/* sprinkle some ships */
 	for(i = 0; i < /*200*/50; i++) {
-		Sprite(SP_SHIP, (int)rnd(de_sitter), (int)rnd(de_sitter), rnd((float)M_PI), scorpion_class, B_STUPID);
+		Sprite(SP_SHIP, rnd(de_sitter), rnd(de_sitter), rnd((float)M_PI), scorpion_class, B_STUPID);
 	}
 	for(i = 0; i < 10; i++) {
-		Sprite(SP_SHIP, (int)rnd(de_sitter), (int)rnd(de_sitter), rnd((float)M_PI), blob_class, B_STUPID);
+		Sprite(SP_SHIP, rnd(de_sitter), rnd(de_sitter), rnd((float)M_PI), blob_class, B_STUPID);
 	}
 
 }
