+++ src/general/Random.c
@@ -27,5 +27,5 @@ int RandomUniformInt(const int max) {
  @param sigma
  @return		Random int about [-max, max]. */
 float RandomUniformFloat(const float max) {
-	return (float)(((double)rand() / (1.0 + RAND_MAX) - 0.5) * 2.0 * (double)max);
+	return (float)((rand() / (1.0 + RAND_MAX) - 0.5) * 2.0 * max);
 }
